+++ tests/lib/DAV_Request_DELETETest.php
@@ -37,6 +37,7 @@ class DAV_Request_DELETETest extends PHPUnit_Framework_TestCase {
     $_SERVER['REQUEST_URI'] = '/path/to/resource';
     DAV::$REGISTRY = new DAV_Test_Registry();
     DAV::$REGISTRY->setResourceClass( 'DAVACL_Test_Resource' );
+    dav::$LOCKPROVIDER = new DAV_Test_Lock_Provider();
     $this->obj = DAV_Request::inst();
   }