+++ CHANGES.md
@@ -1,5 +1,8 @@
-API changes made while creating unit tests
+v1.0.1
+- If you are using Composer to load the library, just run \DAV::bootstrap() otherwise include lib\bootstrap.php (and don't run \DAV::bootstrap() seperately)
+- Set the configuration with \DAV::setDebugFile() instead of the config.ini
 
+API changes made while creating unit tests:
 * To load the library, you need to include /lib/bootstrap.php (instead of /lib/dav.php)
 * DAV::abs2uri() is changed to DAV::path2uri() and now always returns an uri (if you supply a relative path instead of an absolute path, it prefixes the path with the current request URI)
 * DAV::$SUPPORTED_PROPERTIES is now private and can be retrieved by calling DAV::getSupported_Properties()